Improved implementation of QGlib::ValueArray.
The proxy is largely based on the QList/QVector, though there are
still more elements to implement.

It has the downside that operator[] doesn't return a reference
so you can't easily modify values. I did implement a .set(index, value)
operation.

Also I implemented toVector and toList to give the user a way to
work with the collection with real Qt collections.