Implement QGst::Allocator
This provides a partial wrapper for GstAllocator, it allows
finding Allocators registered with glib, unfortunately because
of the difficulties adapting the type signatures for the pointers
to functions needed by glib to create a new GstAllocator you
can't actually directly create an allocator from QGlib.

I inherited from RefCountedObject because allocators we create
need to be deleted when the variable goes out of scope, but
Allocators returned from glib shouldn't be freed. Reusing
the RefCountedObject api made that a little clearer.

The actual implementation still has some issues.
its not thread safe, and dropping the refcount to zero just means
when the Allocator goes out of scope it will free the GstAllocator
pointer hidden in its Private data.